backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-environ==0.11.2
python-telegram-bot==21.6
celery==5.3.6
redis==5.0.1
